refactor(theme): migrate and refactor `genStyleHooks` `genComponentStyleHook` `genSubStyleComponent`.
[RFC](https://github.com/ant-design/ant-design/discussions/49748)
