+++ components/pagination/style/bordered.ts
@@ -3,7 +3,7 @@ import { unit } from '@ant-design/cssinjs';
 import type { PaginationToken } from '.';
 import { prepareComponentToken, prepareToken } from '.';
 import type { GenerateStyle } from '../../theme/interface';
-import { genSubStyleComponent } from '../../theme/util/genComponentStyleHook';
+import { genSubStyleComponent } from '../../theme/internal';
 
 const genBorderedStyle: GenerateStyle<PaginationToken> = (token) => {
   const { componentCls } = token;
